+++ codegen-server/python/src/main/kotlin/software/amazon/smithy/rust/codegen/server/python/smithy/generators/PythonServerRootGenerator.kt
@@ -10,7 +10,7 @@ import software.amazon.smithy.rust.codegen.core.smithy.RustCrate
 import software.amazon.smithy.rust.codegen.core.smithy.generators.protocol.ProtocolSupport
 import software.amazon.smithy.rust.codegen.server.python.smithy.PythonServerRustModule
 import software.amazon.smithy.rust.codegen.server.smithy.ServerCodegenContext
-import software.amazon.smithy.rust.codegen.server.smithy.generators.ServerServiceGenerator
+import software.amazon.smithy.rust.codegen.server.smithy.generators.ServerRootGenerator
 import software.amazon.smithy.rust.codegen.server.smithy.generators.protocol.ServerProtocol
 import software.amazon.smithy.rust.codegen.server.smithy.generators.protocol.ServerProtocolGenerator
 
@@ -20,13 +20,13 @@ import software.amazon.smithy.rust.codegen.server.smithy.generators.protocol.Ser
  * Service generator is the main code generation entry point for Smithy services. Individual structures and unions are
  * generated in codegen visitor, but this class handles all protocol-specific code generation (i.e. operations).
  */
-class PythonServerServiceGenerator(
+class PythonServerRootGenerator(
     private val rustCrate: RustCrate,
     protocolGenerator: ServerProtocolGenerator,
     protocolSupport: ProtocolSupport,
     protocol: ServerProtocol,
     private val context: ServerCodegenContext,
-) : ServerServiceGenerator(rustCrate, protocolGenerator, protocolSupport, protocol, context) {
+) : ServerRootGenerator(rustCrate, protocolGenerator, protocolSupport, protocol, context) {
     override fun renderExtras(operations: List<OperationShape>) {
         rustCrate.withModule(PythonServerRustModule.PythonServerApplication) {
             PythonApplicationGenerator(context, protocol, operations)
